refactor: extract KeyedDecodingContainer+DefaultCodable extension
